Migrate Guest login logic to CQRS pattern.
