+++ src/FoodMaster.WebSite/Models/GuestCredentials.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
+using System.Security.Claims;
+
+using MediatR;
 
 namespace FoodMaster.WebSite.Models
 {
-    public class GuestCredentials
+    public class GuestCredentials : IRequest<Claim[]>
     {
         [Required]
         [DataType(DataType.Text)]
